device: Correct the # of digits to look for in BAR
The way that getDeviceRegion works at present is to run the lspci -s
<device_bdf> and grep for any lines that have size in them. This is to
find the BAR 0 offset from /dev/mem to map and read some MMIO registers.
However, it expects that this offset is always going to be between 11 to
16 digits. That is not the case for certain platforms where the BMG BAR
0 is mapped to an offset that is only 10 digits in size. With this
change, we are now changing the number of digits to search for to be
between 10 and 16.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>