gpu_device_stub: Fix context leak in EU metrics collection in Linux
This commit addresses a context leak issue in the
`toGetEuActiveStallIdleCore` function. The context (`hContext`) was
not being properly destroyed in several error-handling paths, leading to
resource leaks. Key changes include:

- Added calls to `zeContextDestroy` in all error-handling paths to ensure
  the context is properly cleaned up.
- Ensured `zetMetricStreamerClose` is called before destroying the context
  when `zetMetricStreamerReadData` fails.

These changes ensure that the context is always destroyed, preventing
resource leaks and improving the robustness of the EU metrics collection
logic.

This commit was based on this commit which addressed the same issue on
windows:
commit b42d51fdef718b5fdf863af8c368a44df2a4af8d
Author: Zheng Wang <zheng.w.wang@intel.com>
Date:   Wed Feb 5 22:14:49 2025 +0800

    Fix a context leak for EU metrics (#873)