cli: Created a daemon as well from the same code
XPUM is supposed to have two variants. A daemon version and a daemonless
(xpu-smi) version. Both are compiled from the same codebase. Daemon has
a few extra commands that the smi variant doesn't. In this patch, we
also separated the commands based on which ones can be called by the
daemon vs smi version.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>