hal: Added function to find a device by BDF or by index
One of the first things that we need to do is to find a device by the
BDF (for ex: 0000:03:00.0) provided by the user or by the index id (for
ex 0). This patch allows us to do so by checking the BDF in the PCI
class from the stored value for the device.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>