hal: Changed the definition of findDeviceByBDF
Previously, we could only get one device handle and one device from this
function.  With this new function, we now get a list of devices and a
list of device handles. This can be used better in all commands where
the user doesn't provide any device BDF and instead wants a particular
action to be taken upon all devices.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>