xpum: Added a privilige check in the ial and oal layers
XPUM will need to check for whether it has the right priviliges in
daemon mode or not. Since this is a Linux specific thing, it had to be
added in the oal layer with Windows simply returning true. At present,
ial isn't using it for anything. That will be added in another commit at
a later point.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>