hal: Changed the relationship between driver and device are handled
The right way that Sysman enumerates zeDevices and zesDevices is as
follows. Imagine if there are 2 dGPUs (BMG NX2 + BMG B580) connected to
an ARL with 1 iGPU on it. In this case, the number of:
zeDrivers = 2
zesDrivers = 1

zeDrivers[0] represents Xe (for BMG) and zeDrivers[1] represents i915
(for ARL).

zeDrivers[0] has two devices on it:
	zeDevice[0] - representing NX2
	zeDevice[1] - representing B580

zeDrivers[1] has one device on it:
	zeDevice[0] - representing ARL

On the other hand, Sysman enumerates only one driver representing all
devices.

zesDriver[0] has 3 devices on it:
	zesDevice[0] - representing B580
	zesDevice[1] - representing NX2
	zesDevice[2] - ARL

Note that the way that the ze devices are enumerated may not be the same
way as zes devices are enumerated. Meaning, zeDevice[0] may map to
zesDevice[1].

This patch fixes all of these issues and correctly represents these
objects between drivers, device groups and devices.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>