Device: fix all GPU engine utilization
Replace conditional checks for specific engine types
(COMPUTE_SINGLE, RENDER_SINGLE, COPY_SINGLE) with a single check for
ZES_ENGINE_GROUP_ALL in both GPU utilization data handler and device stub.

This change fixes the issue with collecting all GPU engine utilization
metrics.

The reason for doing this change is that previously the code was only
checking for those 3 specific metrics (one was even accidentally
duplicated), and the code was not combining the metrics, it would
only display back out whatever metric happened to get logged last.
Rather than change all the logic to add up all the metrics, and
include all the other engine types that were missing, it is easiest,
and likely best, to just capture the all engine metric which is
already provided.