xpum: Get power for the dump command
Power is defined as Energy used per second. So the way to calculate
power is to find out from Sysman how much energy is being used in
Joules at any given time, then wait for a little bit and get the energy
again. Here we wait for 500 ms. The simply divide the energy delta with
the timestamp delta and get the power used.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>