rest: Setting the response content type as text/plain
The metrics are formatted as a string in exporter.py. Flask
treats all strings as html as a content_type. Grafana needs the
content_type to be text/plain in order to read the metrics.

Configured the Flask response content_type to be text/plain before
returning it to the app.

Signed-off-by: Lui, Jonathan <jonathan.ming.jun.lui@intel.com>