rest: ps: Fixing ps command using RESTful API.
The ps command used xpumGetAllDeviceUtilizationByProcess() and
this function grabs the processes from sysfs that only exists
in i915 dkms. Sysman provides an API to grab the processes and
it is wrapped with xpumGetDeviceProcessStat(). This commit
replaces the xpumGetAllDeviceUtilizationByProcess() with
xpumGetDeviceProcessStat() thus allowing sysman to be the sole
provider of the system processes.

Signed-off-by: Lui, Jonathan <jonathan.ming.jun.lui@intel.com>