xpum: Added support for gpuEnergyConsumed
This required us to change HAL such that we first found out about the
power domain. There are 4 domains that BMG GPUs report. They are card,
package, memory and gpu. We are only interested in card for overall
power and gpu for energy consumed or if this device doesn't have either
card or gpu, then it would have package in which case, we should get
package numbers for both power and energy.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>