cmd_dump: Create thread data at the start of each thread
We can't use static variables inside each function because it will be
called for different devices and we would be comparing old value of one
device with new values of a different device. The right way to do this
comparision is to create a thread data as soon as we create thread args.
This thread data can have the total sum of all values that any dump
function needs and we can use bits and pieces from it.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>