cmd_dump: Added functionality for -n flag
Through this flag, the user can specify that it wants to exit after a
few iterations. Note that if this flag is specified, then we don't honor
a user quitting the funtion through the 'q' keyboard stroke. We expect
that the user wants to run the full set of iterations.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>