ial: Fixed enums being used as array indices
In the current code, enum values were being used as array indices which
requires them to be contiguous and starting from 0. If someone were to
move lines in the corresponding arrays or add an enum in the middle
without adding the corresponding line in the arrays, then this would
cause incorrect indices being accessed.
With this patch, we have changed all arrays in the IAL to
std::unordered_map so that we can still continue to access objects
similar to an array but the actual index location is no longer tied to
the corresponding enums.

This patch fixes Jira #877

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>