xpum: Changed UNUSED to an attribute instead of Macro
The idiom in C++20 would be to use UNUSED as an attribute and not a
macro. Changed all parameters that are currently unused to this
attribute. There were some locally defined variables that also needed to
have UNUSED so created a new macro called UNUSED_VAR for them as we
couldn't use UNUSED anymore.

This fixes Jira #926.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>