cmd_config: Set performance factor
For those platforms that support performance factor (such as ATS-M),
this function will try to set it. For other platforms, it will get a
feature not supported error. The user has to specify which engine do
they want to set the factor for. Currently, only compute and media are
allowed. Factor is a value between 0.0 and 100.0.

This is for XPUM-929.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>