build: Fixed spurious runtime errors that showed up with meson builds
Meson build exposed an issue where sysman related errors were showing up
during XPUM initialization. This was because of multiple copies of
dbgLvl variable (one in the app and another in the dll) that weren't
syncing up. The expected behavior is that during initialization, even if
there are some sysman APIs that aren't working, XPUM should not print
these errors out as the user may not execute those functions at run
time.

This fixes XPUM-951

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>