xpu-smi: Renamed the project to xpu-smi from xpum-smi
Visual Studio project was called xpum-smi and the binary that it was
creating was also of the same name while all other builds (meson for
Windows and Linux) were creating the binary with the right name. With
this patch, we have corrected the name which required the project name
to also be changed.

This fixes XPUM-949

:Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>