firmware: Allocate all types of firmware update classes during init
Instead of dynamically allocating all types of firmware classes every
time a firmware update is called, initialize them at init time. This is
because we need to do a one time initialization in some of these
classes (such as in AMC class). Doing this initialization over and over
again would be redundant and time consuming.

This is part of a fix for XPUM-945.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>