amc: Changed the amcDeviceList to be a list of amcCardInfo
Previously this list just carried string path of the amcDevice. Now, it
is a list of data structure called amcCardInfo which holds both the
string path of the amcDevice but also its gpuParent. The reason for this
change is because we will need to identify which device is associated
with an AMC (and which AMC) at a later point.

This is part of a fix for XPUM-945.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>