updatefw: Update firmware for multiple devices in parallel
Some firmware updates can take several minutes to finish. If the user
has specified that they want multiple cards to be updated, then instead
of serially updating them, create threads for each one and update them
in parallel to speed things up.

This is part of a fix for XPUM-945.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>