amc: Use amcIndex instead of deviceIndex when updating AMC firmware
Not all devices have AMC controller. For example, a system could have
one BMG B60 and another BMG NX2 and only NX2 has an AMC in it. So, while
updating AMC firmware, we need to track which device has AMC on it and
provide to amcupd class only the AMC firmware index otherwise it may
not update the right device.

This is part of a fix for XPUM-945.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>