fwupd: Show progress on a per device basis
This patch finally lets the user see progress on a per device basis. For
example, if 2 AMC devices will be updating their firmware, it will show
something like:
Firmware update progress for device 1: ###### 50 %
Firmware update progress for device 2: #### 40 %

In order to update multiple lines simultaneously, we have to keep track
of the current and total number of threads (lines to print).

This is part of a fix for XPUM-945.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>