cmd_discovery: Find drm device path for a particular device
drm device path can be found by going through /sys/class/drm and looking
for card0/1/... Then going into that directory and finding the device
symlink. The basename $(realpath of this device symlink) is the bdf. If
this bdf matches our card's bdf, then we have found a match and the
corresponding drm device path is /dev/dri/card*. This is of course only
available for Linux so for Windows it will return nothing.

This is part of a fix for XPUM-996.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>