global_lock: Added a cross-process lock while updating firmware
If we are updating the firmware through one instance of XPUM, we can't
allow another instance (process) to try doing another update. This is
due to the fact that for certain firmwares like AMC, XPUM has to
initialize the i2c child device driver and this can interfere with an
ongoing firmware update.
In order to prevent another instance from trying to do a firmware
update, we had to create Windows and Linux based files whose presence
will be checked by each XPUM instance before starting any firmware
updates.

This is part of a fix for XPUM-945.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>