cmd_discovery: Add listamcversions feature
This feature required a tweak to how we call getFWversion as we now have
to pass the GPU BDF. If this is for the AMC firmware, we will use
amcGetCardInfo to get the serialnumber and version info.
This required us to introduce two new functions in amclib class to
find out which AMC device belongs to which GPU and also get FRU data for
that device.

This is part of a fix for XPUM-994.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>