pldm: Relax checks for AMC FW response for mctp and pldm
When we get FRU response, we check for the command type in header to match
the command that we sent. However, we don't get this command back. For
example, we sent a 1 for cmdType but we get back 4. Similarly other PLDM
and MCTP command checks are relaxed to overcome FW issues.

This is part of a fix for XPUM-945.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>