amcupd: Use a static flag for number of AMC cards
Instead of a per class instance flag, we need to use a single static
instance so that we can provide the total number of cards to each class
instance and check if we are going above the limit.

With this commit, we can finally update the AMC FW of multiple cards
simultaneously now!

This is part of a fix for XPUM-945.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>