vgpu: Added support for creating VFs
This patch adds vgpu -c command to XPUM. Through this command, a user
can now create multiple virtual functions. It does several checks to
ensure that these VFs can be created. Also, a vgpu.conf file must be
present in the resources/config folder where XPUM is being run from as
it needs to read it to create the VFs and set appropriate configuration
settings for them.

Related-To: XPUM-924

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>