vgpu: Removed unnecessary addkernelparam and assumyes options
addkernelparam was trying to modify the /etc/default/grub and then
calling update-grub equivalent on a per Linux distro. This was very
cumbersome and unnecessary as the system administrator would already
know how to configure the kernel parameters. Similarly, assumeyes is not
needed as all options act upon user's direction.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>