ci: Enable failure on code formatting
Update CI to now fail if any files modified in the PR are incorrectly
formatted. This will fail the 'code-quality' stage. Previously it was
allowed to continue on failure, but GitHub does not properly alert
developers of this unless they click on 'code-quality' themselves.

By failing the stage we now guarantee that we cannot silently merge
incorrectly formatted code. We can further configure GitHub via
Inventory rules to not block a PR from merging if this stage fails, but
it will alert developers.

Related-To: XPUM-1083

Signed-off-by: Sean Avery <sean.avery@intel.com>