diagnostic: Moved the computation test from device to diagnostic class
All the computation functions that are making Level Zero calls should be
part of the diagnostic class instead of the parent device class as there
is no need for them to be in device class.
There was an additional problem created by a prior commits that made the
computation diag test unable to be executed. This commits had
inadvertantly added new enums when DIAG_COMPUTATION was purposely set to
1 as it was the value that a user would input through the command line.
This patch fixed both of these issues.

Fixes XPUM-1084.

Signed-off-by: Singh, Satyeshwar <satyeshwar.singh@intel.com>