ci: Update code-quality to work on forks
The code-quality branch now requires accessing the origin remote to
perform git-clang-format. This change seems to only work when the PR
branch is in the origin repository, and fails on PRs from forks.

This should update the checkout stage to support forked repos.

Set timeout restriction on clang-tidy to speed up the pipeline, as it
provides no current value either way.

Related-To: XPUM-1083

Signed-off-by: Sean Avery <sean.avery@intel.com>