example/list: dump fields in human-readable format
Add new examples/common package for common helpers and utilities for
examples. Move the dump() function to examples/common and extend it to
better print the level-zero-types: any type that implements fmt.Stringer
interface is printed as string.

With this, we get human-readable output for "string property" (i.e.
[]byte) fields and now also enums are printed as strings instead of
numerical values.